# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Domestic Cattle AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export interface LogoutSuccess {
  success: boolean;
}

export interface BreedPrediction {
  breed: string;
  confidence: number;
  rank: number;
}

export interface Analysis {
  id: number;
  userId: string;
  /** @nullable */
  imageData?: string | null;
  /** @nullable */
  imageName?: string | null;
  topBreed: string;
  topConfidence: number;
  predictions: BreedPrediction[];
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface CreateAnalysisBody {
  /** Base64 encoded image data */
  imageData: string;
  /** @nullable */
  imageName?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface AnalysesStats {
  totalAnalyses: number;
  thisMonthAnalyses: number;
  /** @nullable */
  mostCommonBreed?: string | null;
  avgConfidence: number;
  recentAnalyses: Analysis[];
}

export interface BreedCount {
  breed: string;
  count: number;
}

export interface DeleteResult {
  success: boolean;
}

/**
 * Opaque session token — Bearer <sid>.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type ListAnalysesParams = {
  limit?: number;
  offset?: number;
};