# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Domestic Cattle AI API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
  - name: analyses
    description: Cattle breed analyses
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'

  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.

  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the returnTo path on success.

  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.

  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

  /analyses:
    get:
      operationId: listAnalyses
      tags: [analyses]
      summary: List all analyses for current user
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Analysis"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

    post:
      operationId: createAnalysis
      tags: [analyses]
      summary: Submit a cattle image for breed analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnalysisBody"
      responses:
        "201":
          description: Analysis created with breed predictions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /analyses/stats:
    get:
      operationId: getAnalysesStats
      tags: [analyses]
      summary: Get dashboard statistics for current user
      responses:
        "200":
          description: Analysis statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysesStats"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /analyses/breed-stats:
    get:
      operationId: getBreedStats
      tags: [analyses]
      summary: Get breed frequency statistics for all user analyses
      responses:
        "200":
          description: Breed frequency breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BreedCount"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /analyses/{id}:
    get:
      operationId: getAnalysis
      tags: [analyses]
      summary: Get a single analysis by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

    delete:
      operationId: deleteAnalysis
      tags: [analyses]
      summary: Delete an analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /analyses/{id}/breed-distribution:
    get:
      operationId: getBreedDistribution
      tags: [analyses]
      summary: Get breed distribution for a specific analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Breed frequency breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BreedCount"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — Bearer <sid>.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
          format: email
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]

    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'

    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1

    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string

    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    BreedPrediction:
      type: object
      properties:
        breed:
          type: string
        confidence:
          type: number
        rank:
          type: integer
      required:
        - breed
        - confidence
        - rank

    Analysis:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        imageData:
          type: ["string", "null"]
        imageName:
          type: ["string", "null"]
        topBreed:
          type: string
        topConfidence:
          type: number
        predictions:
          type: array
          items:
            $ref: "#/components/schemas/BreedPrediction"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - topBreed
        - topConfidence
        - predictions
        - createdAt

    CreateAnalysisBody:
      type: object
      properties:
        imageData:
          type: string
          description: Base64 encoded image data
        imageName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - imageData

    AnalysesStats:
      type: object
      properties:
        totalAnalyses:
          type: integer
        thisMonthAnalyses:
          type: integer
        mostCommonBreed:
          type: ["string", "null"]
        avgConfidence:
          type: number
        recentAnalyses:
          type: array
          items:
            $ref: "#/components/schemas/Analysis"
      required:
        - totalAnalyses
        - thisMonthAnalyses
        - avgConfidence
        - recentAnalyses

    BreedCount:
      type: object
      properties:
        breed:
          type: string
        count:
          type: integer
      required:
        - breed
        - count

    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success