# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Domestic Cattle AI API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get the currently authenticated user
 */
export const GetCurrentAuthUserHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — Bearer <sid>."),
});

export const GetCurrentAuthUserResponse = zod.object({
  user: zod.union([
    zod.object({
      id: zod.string(),
      email: zod.string().email().nullable(),
      firstName: zod.string().nullable(),
      lastName: zod.string().nullable(),
      profileImageUrl: zod.string().nullable(),
    }),
    zod.null(),
  ]),
});

/**
 * @summary Start the browser OIDC login flow
 */
export const BeginBrowserLoginQueryParams = zod.object({
  returnTo: zod.coerce.string().optional(),
});

/**
 * @summary Complete the browser OIDC login flow
 */
export const HandleBrowserLoginCallbackQueryParams = zod.object({
  code: zod.coerce.string().optional(),
  state: zod.coerce.string().optional(),
  iss: zod.coerce.string().url().optional(),
});

/**
 * @summary Clear the session and begin OIDC logout
 */
export const LogoutBrowserSessionHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — Bearer <sid>."),
});

/**
 * @summary Exchange a mobile OIDC code for a session token
 */

export const ExchangeMobileAuthorizationCodeBody = zod.object({
  code: zod.string().min(1),
  code_verifier: zod.string().min(1),
  redirect_uri: zod.string().url().min(1),
  state: zod.string().min(1),
  nonce: zod.string().min(1).optional(),
});

export const ExchangeMobileAuthorizationCodeResponse = zod.object({
  token: zod.string(),
});

/**
 * @summary Delete a mobile session token
 */
export const LogoutMobileSessionHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — Bearer <sid>."),
});

export const LogoutMobileSessionResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary List all analyses for current user
 */
export const listAnalysesQueryLimitDefault = 50;
export const listAnalysesQueryOffsetDefault = 0;

export const ListAnalysesQueryParams = zod.object({
  limit: zod.coerce.number().default(listAnalysesQueryLimitDefault),
  offset: zod.coerce.number().default(listAnalysesQueryOffsetDefault),
});

export const ListAnalysesResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  imageData: zod.string().nullish(),
  imageName: zod.string().nullish(),
  topBreed: zod.string(),
  topConfidence: zod.number(),
  predictions: zod.array(
    zod.object({
      breed: zod.string(),
      confidence: zod.number(),
      rank: zod.number(),
    }),
  ),
  notes: zod.string().nullish(),
  createdAt: zod.string(),
});
export const ListAnalysesResponse = zod.array(ListAnalysesResponseItem);

/**
 * @summary Submit a cattle image for breed analysis
 */
export const CreateAnalysisBody = zod.object({
  imageData: zod.string().describe("Base64 encoded image data"),
  imageName: zod.string().nullish(),
  notes: zod.string().nullish(),
});

/**
 * @summary Get dashboard statistics for current user
 */
export const GetAnalysesStatsResponse = zod.object({
  totalAnalyses: zod.number(),
  thisMonthAnalyses: zod.number(),
  mostCommonBreed: zod.string().nullish(),
  avgConfidence: zod.number(),
  recentAnalyses: zod.array(
    zod.object({
      id: zod.number(),
      userId: zod.string(),
      imageData: zod.string().nullish(),
      imageName: zod.string().nullish(),
      topBreed: zod.string(),
      topConfidence: zod.number(),
      predictions: zod.array(
        zod.object({
          breed: zod.string(),
          confidence: zod.number(),
          rank: zod.number(),
        }),
      ),
      notes: zod.string().nullish(),
      createdAt: zod.string(),
    }),
  ),
});

/**
 * @summary Get breed frequency statistics for all user analyses
 */
export const GetBreedStatsResponseItem = zod.object({
  breed: zod.string(),
  count: zod.number(),
});
export const GetBreedStatsResponse = zod.array(GetBreedStatsResponseItem);

/**
 * @summary Get a single analysis by ID
 */
export const GetAnalysisParams = zod.object({
  id: zod.coerce.number(),
});

export const GetAnalysisResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  imageData: zod.string().nullish(),
  imageName: zod.string().nullish(),
  topBreed: zod.string(),
  topConfidence: zod.number(),
  predictions: zod.array(
    zod.object({
      breed: zod.string(),
      confidence: zod.number(),
      rank: zod.number(),
    }),
  ),
  notes: zod.string().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary Delete an analysis
 */
export const DeleteAnalysisParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteAnalysisResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get breed distribution for a specific analysis
 */
export const GetBreedDistributionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetBreedDistributionResponseItem = zod.object({
  breed: zod.string(),
  count: zod.number(),
});
export const GetBreedDistributionResponse = zod.array(
  GetBreedDistributionResponseItem,
);