# lib/api-zod/src/generated/types/analysesStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Domestic Cattle AI API
 * OpenAPI spec version: 0.1.0
 */
import type { Analysis } from "./analysis";

export interface AnalysesStats {
  totalAnalyses: number;
  thisMonthAnalyses: number;
  /** @nullable */
  mostCommonBreed?: string | null;
  avgConfidence: number;
  recentAnalyses: Analysis[];
}