# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Domestic Cattle AI API
 * OpenAPI spec version: 0.1.0
 */
import type { BreedPrediction } from "./breedPrediction";

export interface Analysis {
  id: number;
  userId: string;
  /** @nullable */
  imageData?: string | null;
  /** @nullable */
  imageName?: string | null;
  topBreed: string;
  topConfidence: number;
  predictions: BreedPrediction[];
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}