# lib/api-zod/src/generated/types/breedPrediction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Domestic Cattle AI API
 * OpenAPI spec version: 0.1.0
 */

export interface BreedPrediction {
  breed: string;
  confidence: number;
  rank: number;
}