# lib/api-zod/src/generated/types/createAnalysisBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Domestic Cattle AI API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAnalysisBody {
  /** Base64 encoded image data */
  imageData: string;
  /** @nullable */
  imageName?: string | null;
  /** @nullable */
  notes?: string | null;
}